src/frontend/requirements.txt
streamlit
requests
pandas
pyyaml
streamlit-mic-recorder
elevenlabs
groq
